+++ maui/src/Charts/Series/LineSeries.cs
@@ -463,7 +463,7 @@ namespace Syncfusion.Maui.Toolkit.Charts
                     return true;
                 }
 
-                var dataPoint = FindNearestChartPoint((float)point.X, (float)point.Y);
+                var dataPoint = FindNearestChartPoint(point.X, point.Y);
 
                 if (dataPoint == null || ActualData == null)
                 {
